design system cleanup
