--- app/src/main/java/com/github/odaridavid/weatherapp/designsystem/PermissionComponent.kt
@@ -1,122 +1,0 @@
-package com.github.odaridavid.weatherapp.designsystem
-
-import android.Manifest
-import android.app.Activity
-import android.content.Context
-import android.content.pm.PackageManager
-import androidx.activity.result.ActivityResultLauncher
-import androidx.annotation.StringRes
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.AlertDialog
-import androidx.compose.material.Button
-import androidx.compose.material.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-import androidx.core.content.ContextCompat
-import com.github.odaridavid.weatherapp.R
-
-@Composable
-fun PermissionRationaleDialog(
-    isDialogShown: MutableState<Boolean>,
-    activityPermissionResult: ActivityResultLauncher<String>,
-    showWeatherUI: MutableState<Boolean>
-) {
-    AlertDialog(
-        onDismissRequest = { isDialogShown.value = false },
-        title = {
-            Text(text = stringResource(R.string.location_rationale_title))
-        },
-        text = {
-            Text(text = stringResource(R.string.location_rationale_description))
-        },
-        buttons = {
-            Button(onClick = {
-                isDialogShown.value = false
-                activityPermissionResult.launch(Manifest.permission.ACCESS_COARSE_LOCATION)
-            }) {
-                Text(text = stringResource(R.string.location_rationale_button_grant))
-            }
-            Button(onClick = {
-                isDialogShown.value = false
-                showWeatherUI.value = false
-            }) {
-                Text(text = stringResource(R.string.location_rationale_button_deny))
-            }
-        }
-    )
-}
-
-@Composable
-fun RequiresPermissionsScreen() {
-    InfoScreen(message = R.string.location_no_permission_screen_description)
-}
-
-@Composable
-fun Context.CheckForPermissions(
-    onPermissionGranted: @Composable () -> Unit,
-    onPermissionDenied: @Composable () -> Unit
-) {
-    when (ContextCompat.checkSelfPermission(
-        this,
-        Manifest.permission.ACCESS_COARSE_LOCATION
-    )) {
-        PackageManager.PERMISSION_GRANTED -> {
-            onPermissionGranted()
-        }
-        PackageManager.PERMISSION_DENIED -> {
-            onPermissionDenied()
-        }
-    }
-}
-
-@Composable
-fun Activity.OnPermissionDenied(
-    activityPermissionResult: ActivityResultLauncher<String>,
-) {
-    val showWeatherUI = remember { mutableStateOf(false) }
-    if (shouldShowRequestPermissionRationale(Manifest.permission.ACCESS_COARSE_LOCATION)) {
-        val isDialogShown = remember { mutableStateOf(true) }
-        if (isDialogShown.value) {
-            PermissionRationaleDialog(
-                isDialogShown,
-                activityPermissionResult,
-                showWeatherUI
-            )
-        }
-    } else {
-        activityPermissionResult.launch(Manifest.permission.ACCESS_COARSE_LOCATION)
-    }
-}
-
-@Composable
-fun EnableLocationSettingScreen() {
-    InfoScreen(message = R.string.location_settings_not_enabled)
-}
-
-@Composable
-fun InfoScreen(@StringRes message: Int) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp)
-    ) {
-        Spacer(modifier = Modifier.weight(0.5f))
-        Text(
-            text = stringResource(message),
-            modifier = Modifier
-                .padding(16.dp)
-                .align(Alignment.CenterHorizontally)
-        )
-        Spacer(modifier = Modifier.weight(0.5f))
-    }
-}
-
