+++ app/src/main/java/com/github/odaridavid/weatherapp/designsystem/ProgressScreens.kt
@@ -2,17 +2,24 @@ package com.github.odaridavid.weatherapp.designsystem
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
 import androidx.compose.material.CircularProgressIndicator
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
 
 @Composable
 fun LoadingScreen() {
-    Column(modifier = Modifier.fillMaxWidth()) {
+    Column(modifier = Modifier.fillMaxSize()) {
         Spacer(modifier = Modifier.weight(1.0f))
-        CircularProgressIndicator(modifier = Modifier.align(Alignment.CenterHorizontally))
+        CircularProgressIndicator(
+            modifier = Modifier
+                .padding(16.dp)
+                .align(Alignment.CenterHorizontally)
+        )
         Spacer(modifier = Modifier.weight(1.0f))
     }
 }
